requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pydantic>=2.0.0
graphviz>=0.20.1
python-docx>=1.1.0
matplotlib>=3.7.0
pypowsybl>=1.13.0
